Fix bugs and added fractional model
There is a bug in the old version that I applied parameter
transformation multiple times in the residual and cost function. The
parameters transformation should only be applied in the predict
method, since the residual and cost functions call this method.

The fractional model seems to have simple features of the model
manifold, while also has some complexity due to its non-linearity.