Add a function that estimates the autocorrelation length
Although I used emcee's autocorrelation function here, I also added
an option to be able to transform the chains prior to computing the
autocorrelation length. The motivation behind the transformation is
to reduce the correlation between parameters prior to the autocorrelation
calculation. This is because the autocorrelation calculation assumes
independent parameters, which is not always true.